chore(): Splits demo-app component and module
