+++ demo/src/demo-app.component.html
@@ -1,9 +1,9 @@
-<form [formGroup]="userReactiveForm">
+<form [formGroup]="reactiveForm">
   <label for="rut">SKU</label>
   <input type="text"
     id="rut"
     placeholder="Rut"
-    [formControl]="userReactiveForm.controls['rut']"
+    [formControl]="reactiveForm.controls['rut']"
     plRut
     >
 </form>
