+++ demo/src/demo-app.module.ts
@@ -1,0 +1,27 @@
+import { NgModule } from '@angular/core';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { BrowserModule } from '@angular/platform-browser';
+import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+
+import { Ng2Rut } from '../../src/ng2-rut.module';
+import { RutValidator } from '../../src/rut.validator';
+
+import { DemoAppComponent } from './demo-app.component'
+
+@NgModule({
+  declarations: [
+    DemoAppComponent,
+  ],
+  providers: [RutValidator],
+  imports: [
+    BrowserModule,
+    FormsModule,
+    ReactiveFormsModule,
+    Ng2Rut
+  ],
+  bootstrap: [DemoAppComponent],
+})
+class DemoAppModule { }
+
+platformBrowserDynamic().bootstrapModule(DemoAppModule)
+.catch((err: any) => console.error(err));
